refactor: break down cache structure & builder
