+++ src/main/kotlin/org/bibletranslationtools/fetcher/data/cache/AvailabilityCache.kt
@@ -1,0 +1,5 @@
+package org.bibletranslationtools.fetcher.data.cache
+
+data class AvailabilityCache(
+    var languages: List<LanguageCache>
+)
