+++ src/main/kotlin/org/bibletranslationtools/fetcher/data/cache/ChapterCache.kt
@@ -1,0 +1,7 @@
+package org.bibletranslationtools.fetcher.data.cache
+
+data class ChapterCache(
+    val number: Int,
+    var availability: Boolean = false,
+    var url: String? = null
+)
