+++ src/main/kotlin/org/bibletranslationtools/fetcher/data/cache/LanguageCache.kt
@@ -1,0 +1,7 @@
+package org.bibletranslationtools.fetcher.data.cache
+
+data class LanguageCache(
+    val code: String,
+    var availability: Boolean = false,
+    val products: List<ProductCache> = listOf()
+)
