+++ src/main/kotlin/org/bibletranslationtools/fetcher/data/cache/ProductCache.kt
@@ -1,0 +1,7 @@
+package org.bibletranslationtools.fetcher.data.cache
+
+data class ProductCache(
+    val slug: String,
+    var availability: Boolean = false,
+    val books: List<BookCache> = listOf()
+)
