+++ src/main/kotlin/org/bibletranslationtools/fetcher/impl/repository/AvailabilityCacheRepo.kt
@@ -1,0 +1,79 @@
+package org.bibletranslationtools.fetcher.impl.repository
+
+import org.bibletranslationtools.fetcher.data.cache.AvailabilityCache
+import org.bibletranslationtools.fetcher.repository.ContentCacheRepository
+
+class AvailabilityCacheRepo(
+    private val cacheBuilder: ContentAvailabilityCacheBuilder
+) : ContentCacheRepository {
+
+    private var cache: AvailabilityCache
+
+    init {
+        cache = cacheBuilder.build()
+    }
+
+    @Synchronized
+    override fun update() {
+        cache = cacheBuilder.build()
+    }
+
+    override fun isLanguageAvailable(code: String) = cache.languages.any { it.code == code && it.availability }
+
+    override fun isProductAvailable(productSlug: String, languageCode: String): Boolean {
+        return cache.languages.find {
+            it.code == languageCode && it.availability
+        }?.products?.any {
+            it.slug == productSlug
+        } ?: false
+    }
+
+    override fun isBookAvailable(
+        bookSlug: String,
+        languageCode: String,
+        productSlug: String
+    ): Boolean {
+        val productCache = cache.languages.find {
+            it.code == languageCode && it.availability
+        }?.products?.find {
+            it.slug == productSlug && it.availability
+        } ?: return false
+
+        return productCache.books.any {
+            it.slug == bookSlug && it.availability
+        }
+    }
+
+    override fun getChapterUrl(
+        number: Int,
+        bookSlug: String,
+        languageCode: String,
+        productSlug: String
+    ): String? {
+        val bookCache = cache.languages.find {
+            it.code == languageCode && it.availability
+        }?.products?.find {
+            it.slug == productSlug && it.availability
+        }?.books?.find {
+            it.slug == bookSlug && it.availability
+        } ?: return null
+
+        return bookCache.chapters.find {
+            it.number == number
+        }?.url
+    }
+
+    override fun getBookUrl(
+        bookSlug: String,
+        languageCode: String,
+        productSlug: String
+    ): String? {
+        return cache.languages.find {
+            it.code == languageCode && it.availability
+        }?.products?.find {
+            it.slug == productSlug && it.availability
+        }?.books?.find {
+            it.slug == bookSlug
+        }?.url
+    }
+}
