+++ src/main/kotlin/org/bibletranslationtools/fetcher/impl/repository/ContentAvailabilityCacheBuilder.kt
@@ -1,14 +1,16 @@
 package org.bibletranslationtools.fetcher.impl.repository
 
 import io.ktor.http.HttpStatusCode
-import org.bibletranslationtools.fetcher.repository.BookRepository
-import org.bibletranslationtools.fetcher.repository.ChapterCatalog
-import org.bibletranslationtools.fetcher.repository.ContentCacheRepository
-import org.bibletranslationtools.fetcher.repository.DirectoryProvider
+import org.bibletranslationtools.fetcher.data.cache.AvailabilityCache
+import org.bibletranslationtools.fetcher.data.cache.BookCache
+import org.bibletranslationtools.fetcher.data.cache.ChapterCache
+import org.bibletranslationtools.fetcher.data.cache.LanguageCache
+import org.bibletranslationtools.fetcher.data.cache.ProductCache
 import org.bibletranslationtools.fetcher.repository.LanguageCatalog
+import org.bibletranslationtools.fetcher.repository.ChapterCatalog
+import org.bibletranslationtools.fetcher.repository.BookRepository
 import org.bibletranslationtools.fetcher.repository.StorageAccess
-import java.net.HttpURLConnection
-import java.net.URL
+import org.bibletranslationtools.fetcher.repository.DirectoryProvider
 import org.bibletranslationtools.fetcher.usecase.FetchBookViewData
 import org.bibletranslationtools.fetcher.usecase.FetchChapterViewData
 import org.bibletranslationtools.fetcher.usecase.ProductFileExtension
@@ -16,116 +18,28 @@ import org.bibletranslationtools.fetcher.usecase.RequestResourceContainer
 import org.slf4j.LoggerFactory
 import org.wycliffeassociates.resourcecontainer.ResourceContainer
 import java.io.IOException
+import java.net.HttpURLConnection
+import java.net.URL
 
-class ContentAvailabilityCache(
+class ContentAvailabilityCacheBuilder(
     private val languageCatalog: LanguageCatalog,
     private val chapterCatalog: ChapterCatalog,
     private val bookRepository: BookRepository,
     private val storageAccess: StorageAccess,
     directoryProvider: DirectoryProvider
-) : ContentCacheRepository {
+) {
     private val repoDir = directoryProvider.getRCRepositoriesDir()
     private val templateRCName = "%s_%s.zip"
     private val resourceId = "ulb"
-    private var root: List<LanguageCache>
     private val logger = LoggerFactory.getLogger(javaClass)
 
-    private data class LanguageCache(
-        val code: String,
-        var availability: Boolean = false,
-        val products: List<ProductCache> = listOf()
-    )
-
-    private data class ProductCache(
-        val slug: String,
-        var availability: Boolean = false,
-        val books: List<BookCache> = listOf()
-    )
-
-    private data class BookCache(
-        val slug: String,
-        var availability: Boolean = false,
-        val url: String? = null,
-        val chapters: List<ChapterCache> = listOf()
-    )
-
-    private data class ChapterCache(
-        val number: Int,
-        var availability: Boolean = false,
-        var url: String? = null
-    )
-
-    init {
-        root = cacheLanguages()
-    }
-
-    @Synchronized
-    override fun update() {
-        root = cacheLanguages()
-    }
-
-    override fun isLanguageAvailable(code: String) = root.any { it.code == code && it.availability }
-
-    override fun isProductAvailable(productSlug: String, languageCode: String): Boolean {
-        return root.find {
-            it.code == languageCode && it.availability
-        }?.products?.any {
-            it.slug == productSlug
-        } ?: false
-    }
-
-    override fun isBookAvailable(
-        bookSlug: String,
-        languageCode: String,
-        productSlug: String
-    ): Boolean {
-        val productCache = root.find {
-            it.code == languageCode && it.availability
-        }?.products?.find {
-            it.slug == productSlug && it.availability
-        } ?: return false
-
-        return productCache.books.any {
-            it.slug == bookSlug && it.availability
-        }
-    }
-
-    override fun getChapterUrl(
-        number: Int,
-        bookSlug: String,
-        languageCode: String,
-        productSlug: String
-    ): String? {
-        val bookCache = root.find {
-            it.code == languageCode && it.availability
-        }?.products?.find {
-            it.slug == productSlug && it.availability
-        }?.books?.find {
-            it.slug == bookSlug && it.availability
-        } ?: return null
-
-        return bookCache.chapters.find {
-            it.number == number
-        }?.url
-    }
-
-    override fun getBookUrl(
-        bookSlug: String,
-        languageCode: String,
-        productSlug: String
-    ): String? {
-        return root.find {
-            it.code == languageCode && it.availability
-        }?.products?.find {
-            it.slug == productSlug && it.availability
-        }?.books?.find {
-            it.slug == bookSlug
-        }?.url
+    fun build(): AvailabilityCache {
+        return AvailabilityCache(cacheLanguages())
     }
 
     private fun cacheLanguages(): List<LanguageCache> {
         val glList = languageCatalog.getAll()//.filter {it.code == "en" ||it.code == "fr"}
-            return glList.map { lang ->
+        return glList.map { lang ->
             println(glList.indexOf(lang))
             val products = cacheProducts(lang.code)
             val isAvailable = products.any { it.availability }
